flaskbackend/requirements.txt
langchain
langchain-community
sentence-transformers
transformers
# torch==2.7.0 --index-url https://download.pytorch.org/whl/cpu
keras
tf-keras
beautifulsoup4
pdfplumber
flask
google-generativeai
faiss-cpu
python-dotenv